Heating Recipe
-Add ability to accept json as a ISP method
-Several minor fixes

+++ src/main/java/com/notenoughmail/kubejs_tfc/recipe/LoomRecipeJS.java
@@ -25,7 +25,7 @@ public class LoomRecipeJS extends TFCRecipeJS {
 
         steps = count;
         if (listJS.size() > 2) {
-            steps = Integer.parseInt(listJS.get(2).toString());
+            steps = (int) Float.parseFloat(listJS.get(2).toString());
         }
     }
